Refactor: split junit tests into smaller classes
